Performance: improve import logic
After investigating how the regular Python importlib machinery discovers imports (let's call that vanilla), and how
astroid does it, it looks like astroid is slightly less eager to see a directory as a namespace package than vanilla
behaviour.

Rather than trying to resolve every import by looking at every entry in `sys.path` as a regular package and then as a
namespace package like vanilla, it looks at every entry as a regular package and then goes every entry again to look at
it like a namespace package. This saves astroid from having to apply some additional namespace-grouping logic after the
import is found like vanilla behaviour has to do.

However, this also means that it takes slightly longer to resolve every import when so many of them are a namespace
package, which is blown up by how many times it goes through this logic for every import (once to check whether it's a
standard library module, once to check where the import is actually defined...).

This adds an extra condition earlier in the package discovery logic to recognize a directory as a namespace package.

Co-authored-by: Tim Geypens <tim@wayve.ai>